backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==1.10.13
python-dotenv==1.0.1
google-generativeai